Renaming cudnn_conv_runner to gpu_conv_runner
